transcription/requirements.txt
azure-storage-blob
azure-cognitiveservices-speech
python-dotenv
openai
requests